Reboot Cacher

+++ core/class/system/cache.php
@@ -16,34 +16,38 @@ die();
 
 final class cache {
 
-	private static $type = "file";
+	private static $type = CACHE_NONE;
 	private static $connect = false;
 	private static $live_time = 2592000;
+	private static $conn_link = null;
+	private static $conn_path = null;
 
 	public function cache() {
 	global $config;
-		if($config['cache']['activ']) {
-			if((class_exists("Memcached") && $config['cache']['type'] == 2) || (class_exists("Memcached") && $config['cache']['type'] == 1)) {
-				self::$type = "memcached";
-				self::$connect = new Memcached();
-				self::$connect->addServer($config['cache']['server'], $config['cache']['port']) or die ("Could not connect");
-			} elseif((class_exists('Memcache') && $config['cache']['type'] == 2) || (class_exists("Memcache") && $config['cache']['type'] == 2)) {
-				self::$type = "memcache";
-				self::$connect = new Memcache();
-				self::$connect->addServer($config['cache']['server'], $config['cache']['port']) or die ("Could not connect");
-			}
-		} else {
-			self::$type="off";
+		self::$type = $config['cache']['type'];
+		self::$conn_path = $config['cache']['path'];
+		if(class_exists("Memcached") && $config['cache']['type'] == CACHE_MEMCACHED) {
+			self::$connect = new Memcached();
+			self::$connect->addServer($config['cache']['server'], $config['cache']['port']) or die ("Could not connect");
+		} elseif(class_exists('Memcache') && $config['cache']['type'] == CACHE_MEMCACHE) {
+			self::$connect = new Memcache();
+			self::$connect->addServer($config['cache']['server'], $config['cache']['port']) or die ("Could not connect");
+		} elseif(self::$type == CACHE_FTP) {
+			self::$connect = ftp_connect($config['cache']['server'], $config['cache']['port']);
+			ftp_login(self::$connect, $config['cache']['login'], $config['cache']['pass']);
+			self::$conn_link = "ftp://".$config['cache']['login'].":".$config['cache']['pass']."@".$config['cache']['server'].":".$config['cache']['port'].self::$conn_path;
 		}
 	}
 
 	public static function Mtime($data) {
-		if(self::Exists($data) && self::$type!="off") {
-			if(self::$type !== "file") {
+		if(self::Exists($data)) {
+			if(self::$type == CACHE_MEMCACHE || self::$type == CACHE_MEMCACHED) {
 				$data = self::$connect->get($data);
 				return $data['time'];
-			} else {
+			} elseif(self::$type == CACHE_FILE) {
 				return filemtime(ROOT_PATH."core/cache/".$data.".txt");
+			} elseif(self::$type == CACHE_FTP) {
+				return ftp_mdtm(self::$connect, self::$conn_path.$data.".txt");
 			}
 		} else {
 			return 0;
@@ -51,16 +55,18 @@ final class cache {
 	}
 
 	public static function Get($data) {
-		if(self::Exists($data) && self::$type!="off") {
-			if(self::$type !== "file") {
+		if(self::Exists($data)) {
+			if(self::$type == CACHE_MEMCACHE || self::$type == CACHE_MEMCACHED) {
 				$data = self::$connect->get($data);
 				return $data['data'];
-			} else {
+			} elseif(self::$type == CACHE_FILE) {
 				if(file_exists(ROOT_PATH."core/cache/".$data.".txt")) {
 					return unserialize(file_get_contents(ROOT_PATH."core/cache/".$data.".txt"));
 				} else {
 					return false;
 				}
+			} elseif(self::$type == CACHE_FTP) {
+				return unserialize(file_get_contents(self::$conn_link.$data.".txt"));
 			}
 		} else {
 			return false;
@@ -72,35 +78,37 @@ final class cache {
 	}
 
 	public static function Exists($data) {
-		if(self::$type!="off") {
-			if(self::$type !== "file") {
-				if(@(self::$connect->get($data))) {
-						return true;
-				} else {
-					return false;
-				}
+		if(self::$type == CACHE_MEMCACHE || self::$type == CACHE_MEMCACHED) {
+			if(@(self::$connect->get($data))) {
+					return true;
 			} else {
-				return file_exists(ROOT_PATH."core/cache/".$data.".txt");
+				return false;
 			}
+		} elseif(self::$type == CACHE_FILE) {
+			return file_exists(ROOT_PATH."core/cache/".$data.".txt");
+		} elseif(self::$type == CACHE_FTP) {
+			return (ftp_size(self::$connect, self::$conn_path.$data.".txt")>0);
 		}
 	}
 
 	public static function Set($name, $val) {
-		if(self::$type!="off") {
-			if(self::$type !== "file") {
-				return self::$connect->set($name, array("time" => time(), "data" => $val), MEMCACHE_COMPRESSED, self::$live_time);
-			} else {
-				return file_put_contents(ROOT_PATH."core/cache/".$name.".txt", serialize($val));
-			}
+		if(self::$type == CACHE_MEMCACHE || self::$type == CACHE_MEMCACHED) {
+			return self::$connect->set($name, array("time" => time(), "data" => $val), MEMCACHE_COMPRESSED, self::$live_time);
+		} elseif(self::$type == CACHE_FILE) {
+			return file_put_contents(ROOT_PATH."core/cache/".$name.".txt", serialize($val));
+		} elseif(self::$type == CACHE_FTP) {
+			return file_put_contents(self::$conn_link.$name.".txt", serialize($val), 0, stream_context_create(array('ftp' => array('overwrite' => true))));
 		}
 	}
 
 	public static function Delete($name) {
-		if(self::$type!="off" && self::Exists($name)) {
-			if(self::$type !== "file") {
+		if(self::Exists($name)) {
+			if(self::$type == CACHE_MEMCACHE || self::$type == CACHE_MEMCACHED) {
 				return self::$connect->delete($name);
-			} else if(file_exists(ROOT_PATH."core/cache/".$name.".txt") && !is_dir(ROOT_PATH.'core/cache/'.$name.".txt")) {
+			} else if(self::$type == CACHE_FILE && file_exists(ROOT_PATH."core/cache/".$name.".txt") && !is_dir(ROOT_PATH.'core/cache/'.$name.".txt")) {
 				return unlink(ROOT_PATH."core/cache/".$name.".txt");
+			} elseif(self::$type == CACHE_FTP) {
+				return ftp_delete(self::$connect, self::$conn_path.$name.".txt");
 			}
 		} else {
 			return false;
@@ -108,7 +116,7 @@ final class cache {
 	}
 
 	public static function Clear_cache($cache_areas = false) {
-		if(self::$type !== "file") {
+		if(self::$type == CACHE_MEMCACHE || self::$type == CACHE_MEMCACHED) {
 			self::$connect->flush();
 		}
 
@@ -133,10 +141,12 @@ final class cache {
 	}
 
 	public function __destruct() {
-		if(self::$type === "memcached") {
+		if(self::$type == CACHE_MEMCACHED) {
 			self::$connect->quit();
-		} elseif(self::$type === "memcache") {
+		} elseif(self::$type == CACHE_MEMCACHE) {
 			self::$connect->close();
+		} elseif(self::$type == CACHE_FTP) {
+			ftp_close(self::$connect);
 		}
 	}
